Move libraries to lib
