+++ src/elements/emby-tabs/emby-tabs.js
@@ -1,6 +1,6 @@
 import 'webcomponents.js/webcomponents-lite';
 import dom from '../../scripts/dom';
-import ScrollerFactory from '../../libraries/scroller';
+import ScrollerFactory from 'lib/scroller';
 import browser from '../../scripts/browser';
 import focusManager from '../../components/focusManager';
 import layoutManager from '../../components/layoutManager';
