+++ src/lib/scroller.js
